Optimize solver reset(), replaceAll() (#26)
* Optimize solver reset()

For whatever reason, ZSolver::reset() is a very costly operation; this
may be due to some bug in Z3. In the meantime, however, an alternative
implementation using ZSolver::push() and pop() has been provided.
The semantics are unchanged, but execution is considerably faster.

* Optimize Looped replaceAll() Calls

Several places in all versions of RndLearner were calling
replaceAll() in a loop (for each variable) instead of using the vector
version. These have been replaced to use the vector version, which
results in a noticable speed-up (especially for V1, which was using them
in the process of checking lemmas).

* Optimize isSat()

SMTUtils::isSat() was performing an expensive filter()
operation on every invocation; this has been moved into the location
it's actually needed, getModel().