Fix BndExpl For Nested Array Operations
Some benchmarks include nested array operations, e.g.
(store (store ...) j (select ...), which leads to models with the
reverse nesting, e.g. (select (store ...) j); some code in BndExpl
wasn't detecting this properly, leading to a Segmentation Fault as
it blindly assumed the array variable was in the first position.
This has been fixed.

Also, V3 was assuming the second argument of a quantifier was the body;
this has been fixed (reproduce by only including the change to BndExpl
on the below benchmark).

To reproduce:
    freqhorn --v3 --disj --all-mbp --stren-mbp \
        bench_horn_multiple/array_bubble_sort.smt2